requirements.txt
python-dotenv
chromadb
langchain_huggingface
langchain_community
langchain-groq
langchain_chroma
pypdf
langgraph